requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
waitress==2.1.2
requests==2.31.0
resend==0.8.0
